// src/typings/idMap.ts
export enum FooterIdMap {
  RECEPTION = 'footer-reception-btn',
  MANAGEMENT = 'footer-management-btn',
  COMMUNICATION = 'footer-communication-btn',
  COOPERATION = 'footer-cooperation-btn',
  DASHBOARD = 'footer-dashboard-btn',
  PRICE = 'footer-price-btn',
  UPDATE = 'footer-update-btn',
  FAQ = 'footer-faq-btn',
  GUIDE = 'footer-guide-btn',
  LEAFLET = 'footer-leaflet-btn',
  ASK = 'footer-ask-btn',
  NAVERBLOG = 'footer-naver-blog-btn',
}

export enum NavIdMap {
  RECEPTION = 'tnb-reception-btn',
  MANAGEMENT = 'tnb-management-btn',
  COMMUNICATION = 'tnb-communication-btn',
  COOPERATION = 'tnb-cooperation-btn',
  DASHBOARD = 'tnb-dashboard-btn',
  PRICE = 'tnb-price-btn',
  UPDATE = 'tnb-update-btn',
  FAQ = 'tnb-faq-btn',
  GUIDE = 'tnb-guide-btn',
  LEAFLET = 'tnb-leaflet-btn',
  ASK = 'tnb-ask-btn',
  LOGIN = 'tnb-login-btn',
  SIGNUP = 'tnb-start-btn',
}

export enum RegisterIdMap {
  SIGNUP = 'modal-start-btn',
  MAINSIGNUP = 'main-signup-btn',
  LOGIN = 'modal-login-btn',
}

export enum NextFunctionBtnIdMap {
  TO_MANAGEMENT = 'reception-management-btn',
  TO_COMMUNICATION = 'management-communication-btn',
  TO_COOPERATION = 'communication-cooperation-btn',
  TO_DASHBOARD = 'cooperation-dashboard-btn',
}

export enum AskBtnIdMap {
  ASK = 'ask-complete-btn',
}

export enum CtaBtnIdMap {
  ASK = 'remind-ask-btn',
  LOGIN = 'remind-start-btn',
}